fix: adjust applier versions
- V2 applier is updated to be the same as older versions.
- V3 applier is updated to contain sequence flow migrations
- V4 applier is removed as it is not needed anymore
