+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/appliers/EventAppliers.java
@@ -253,11 +253,6 @@ public final class EventAppliers implements EventApplier {
         new ProcessInstanceElementMigratedV3Applier(
             elementInstanceState, processState, state.getMessageState()));
     register(
-        ProcessInstanceIntent.ELEMENT_MIGRATED,
-        4,
-        new ProcessInstanceElementMigratedV4Applier(
-            elementInstanceState, processState, state.getMessageState()));
-    register(
         ProcessInstanceIntent.ANCESTOR_MIGRATED,
         new ProcessInstanceAncestorMigratedApplier(elementInstanceState));
   }
