+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/appliers/ProcessInstanceElementMigratedV2Applier.java
@@ -8,7 +8,6 @@
 package io.camunda.zeebe.engine.state.appliers;
 
 import io.camunda.zeebe.engine.processing.deployment.model.element.ExecutableFlowElementContainer;
-import io.camunda.zeebe.engine.processing.deployment.model.element.ExecutableSequenceFlow;
 import io.camunda.zeebe.engine.state.TypedEventApplier;
 import io.camunda.zeebe.engine.state.immutable.ProcessState;
 import io.camunda.zeebe.engine.state.mutable.MutableElementInstanceState;
@@ -18,7 +17,6 @@ import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.camunda.zeebe.protocol.record.value.BpmnElementType;
 import io.camunda.zeebe.util.buffer.BufferUtil;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.stream.IntStream;
 import org.agrona.DirectBuffer;
 
 /** Applies state changes for `ProcessInstance:Element_Migrated` */
@@ -40,11 +38,6 @@ final class ProcessInstanceElementMigratedV2Applier
 
   @Override
   public void applyState(final long elementInstanceKey, final ProcessInstanceRecord value) {
-    if (value.getBpmnElementType() == BpmnElementType.SEQUENCE_FLOW) {
-      migrateTakenSequenceFlow(value);
-      return;
-    }
-
     if (value.getBpmnElementType() == BpmnElementType.PROCESS) {
       migrateCorrelatedMessageStartEvent(elementInstanceKey, value);
     }
@@ -69,34 +62,6 @@ final class ProcessInstanceElementMigratedV2Applier
       elementInstanceState.insertProcessInstanceKeyByDefinitionKey(
           value.getProcessInstanceKey(), value.getProcessDefinitionKey());
     }
-  }
-
-  private void migrateTakenSequenceFlow(final ProcessInstanceRecord sequenceFlowRecord) {
-    final ExecutableSequenceFlow sequenceFlow =
-        processState.getFlowElement(
-            sequenceFlowRecord.getProcessDefinitionKey(),
-            sequenceFlowRecord.getTenantId(),
-            sequenceFlowRecord.getElementIdBuffer(),
-            ExecutableSequenceFlow.class);
-    final var migratedSequenceFlowId = sequenceFlow.getId();
-    final var targetGatewayId = sequenceFlow.getTarget().getId();
-
-    elementInstanceState.visitTakenSequenceFlows(
-        sequenceFlowRecord.getFlowScopeKey(),
-        (flowScopeKey, sourceGatewayId, sequenceFlowId, number) -> {
-          if (BufferUtil.equals(migratedSequenceFlowId, sequenceFlowId)) {
-            IntStream.range(0, number)
-                .forEach(
-                    ignore -> {
-                      // decrement the number of taken sequence flows for the old gateway
-                      elementInstanceState.decrementNumberOfTakenSequenceFlows(
-                          flowScopeKey, sourceGatewayId, sequenceFlowId);
-                      // increment the number of taken sequence flows for the new gateway
-                      elementInstanceState.incrementNumberOfTakenSequenceFlows(
-                          flowScopeKey, targetGatewayId, sequenceFlowId);
-                    });
-          }
-        });
   }
 
   private void migrateCorrelatedMessageStartEvent(
